tweak: command error handling

+++ Code/client/Events/SetTimeCommandEvent.h
@@ -8,6 +8,6 @@ struct SetTimeCommandEvent
     {
     }
 
-    int32_t Hours;
-    int32_t Minutes;
+    uint8_t Hours;
+    uint8_t Minutes;
 };
